refactor: parse config file contents into a struct
Before this PR, we read every field in the config file individually, which made the code complex. In this PR, we refactor our code to unmarshall the config file into a struct. When then use this config struct to setup the audit client and interceptor.

The struct was created in a previous PR: #55

Additionally, we rename filter to condition to match the Java implementation.