chore: update go client to use pkg/gcputil
The cloud logging processor relies on the metadta server to resolve the project id if no logging client is provided. This causes issues with local testing of services that use the go client. Switched to the gcputil ProjectID function in pkg which checks the environment first and then falls back on the metadata server.